pyproject.toml
[tool.poetry]
name = "bonostyres"
version = "0.1.0"
description = "F1 RESTful API"
authors = ["reardek <jjkowalinski@yahoo.com>"]
license = "MIT"

[tool.poetry.dependencies]
python = "^3.10"
fastapi = "^0.70.0"
SQLAlchemy = "^1.4.27"
python-dotenv = "^0.19.2"
psycopg2-binary = "^2.9.2"
uvicorn = "^0.15.0"
asyncpg = "^0.25.0"

[tool.poetry.dev-dependencies]
autopep8 = "^1.6.0"
alembic = "^1.7.5"

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"
